WIP, ENH: support frame bounds for movie renders
* Fixes #15

* Add a new `frame_bounds` argument to `render()`, so that the start
and end frame may be controlled conveniently by the user when producing
a movie. I suppose the argument for why this should be available to
the user vs. using raw `bpy` would be similar to the argument in favor
of the original `frame` argument to render individual frames.

* I can add some regression tests, but I'll wait to see if folks
actually want this. The current testsuite still passes locally at least.