WIP, ENH: user-contributed material demo
* This is perhaps a bit of a stretch, but it attempts to demonstrate
a crude approach to user-contributed materials. This seems to be quite
different from the current approach of using MN materials loaded from
a binary `.blend` file upstream.

* I wasn't able to get the coloration to work with the custom material
(because of deviation from the node structure used/expected by MN?). One
advantage to having a script/text based approach to material generation
is that provenance is made quite clear and version controlled vs.
needing to extract from a binary file. That said, it may be largely
impractical to do in `bpy`, not sure.

[ci skip]