TST: use open_mainfile()
* It seems that using `open_mainfile()` rather than
`read_homefile` in the `_clean_up()` function is more robust
for our testsuite (doesn't cause failures for
`test_show_trajectory_with_material` for example). One point
of caution is that running locally with multiple threads with
`pytest-xdist` does seem to run into issues with this approach,
perhaps because of the reading of `bpy.data.filepath`.